+++ src/api/endpoints/collections/get-collection/v2.ts
@@ -166,6 +166,7 @@ export const getCollectionV2Options: RouteOptions = {
           ARRAY(
             SELECT "t"."image" FROM "tokens" "t"
             WHERE "t"."collection_id" = "c"."id"
+            AND "t"."image" IS NOT NULL
             LIMIT 4
           ) AS "sample_images"          
         FROM "collections" "c"
@@ -227,7 +228,7 @@ export const getCollectionV2Options: RouteOptions = {
           FROM "token_sets" "ts"
           LEFT JOIN "orders" "o"
             ON "ts"."top_buy_id" = "o"."id"
-          WHERE "ts"."id" = "x"."token_set_id"
+          WHERE "ts"."collection_id" = "x"."id"
           ORDER BY "ts"."top_buy_value" DESC NULLS LAST
           LIMIT 1
         ) "z" ON TRUE