+++ src/api/endpoints/collections/get-collections/v4.ts
@@ -148,6 +148,7 @@ export const getCollectionsV4Options: RouteOptions = {
             SELECT array(
               SELECT tokens.image FROM tokens
               WHERE tokens.collection_id = collections.id
+              AND tokens.image IS NOT NULL
               LIMIT 4
             )
           ) AS sample_images,