+++ src/api/endpoints/orders/get-order-executed/v1.ts
@@ -19,8 +19,7 @@ export const getOrderExecutedV1Options: RouteOptions = {
   },
   validate: {
     query: Joi.object({
-      id: Joi.string().required()
-      ,
+      id: Joi.string().required(),
       checkRecentEvents: Joi.boolean().default(false),
     }),
   },