+++ src/api/endpoints/tokens/get-tokens-bootstrap/v1.ts
@@ -40,14 +40,15 @@ export const getTokensBootstrapV1Options: RouteOptions = {
         .description(
           "Filter to a particular contract. Example: `0x8d04a8c79ceb0889bdd12acdf3fa9d207ed3ff63`"
         ),
-      continuation: Joi.string().pattern(base64Regex)
-      .description(
-        "Use continuation token to request next offset of items."
-      ),
-      limit: Joi.number().integer().min(1).max(500).default(500)
-      .description(
-        "Amount of items returned in response."
-      ),
+      continuation: Joi.string()
+        .pattern(base64Regex)
+        .description("Use continuation token to request next offset of items."),
+      limit: Joi.number()
+        .integer()
+        .min(1)
+        .max(500)
+        .default(500)
+        .description("Amount of items returned in response."),
     })
       .or("collection", "contract")
       .oxor("collection", "contract"),